refactor(Major refactor):OS, Applications, Windows, layouts
This commit introduced some issues that need to be fixed. It was intended to make a partial commit, because the code was quite consequent. The whole application's logic had to be thrown apart, Applications are now stored in a context, and use reducers internaly and Applications have their own class accordingly.